requirements.txt
# Flask
Flask
Flask-SQLAlchemy
Flask-JWT-Extended

# Database
SQLAlchemy
psycopg2-binary
redis==5.1.1

# Others
python-dotenv
